Many code refactors for Dataset to be split into MutableDataset and StreamingDataset. Also generic methods like get_dataset, get_user and get_project are now avaiable directly from the main scrunch module. Removed redundancy in Project.get_dataset method as the get_dataset method now allows to search in projects. also get_dataset method got split into get_streaming_dataset and includes validation that the user is trying to retrieve the correct kind of dataset by the method.
