+++ scrunch/expressions.py
@@ -698,7 +698,7 @@ def prettify(expr, ds=None):
 
         if not is_url:
             return var
-        elif not isinstance(ds, scrunch.datasets.Dataset):
+        elif not isinstance(ds, scrunch.datasets.BaseDataset):
             raise Exception(
                 'Valid Dataset instance is required to resolve variable urls '
                 'in the expression'
