+++ scrunch/tests/test_recodes.py
@@ -2,7 +2,8 @@ import mock
 from unittest import TestCase
 
 import pytest
-from scrunch.datasets import Dataset, Variable
+from scrunch.datasets import Variable
+from scrunch.mutable_dataset import MutableDataset
 from scrunch.variables import responses_from_map
 from scrunch.helpers import subvar_alias
 
@@ -70,22 +71,22 @@ RECODES_PAYLOAD = {
 }
 
 COMBINE_RESPONSES_PAYLOAD = {
-            'element': 'shoji:entity',
-            'body': {
-                'name': 'name',
-                'description': '',
-                'alias': 'alias',
-                'derivation': {
-                    'function': 'combine_responses',
-                    'args': [
-                        {'variable': var_url},
-                        {'value': [
-                            {'alias': 'alias_1', 'combined_ids': [subvar1_url, subvar2_url], 'name': 'online'}
-                        ]}
-                    ]
-                }
-            }
+    'element': 'shoji:entity',
+    'body': {
+        'name': 'name',
+        'description': '',
+        'alias': 'alias',
+        'derivation': {
+            'function': 'combine_responses',
+            'args': [
+                {'variable': var_url},
+                {'value': [
+                    {'alias': 'alias_1', 'combined_ids': [subvar1_url, subvar2_url], 'name': 'online'}
+                ]}
+            ]
         }
+    }
+}
 
 
 class TestCombine(TestCase):
@@ -139,7 +140,7 @@ class TestCombine(TestCase):
         resource.variables.by.return_value = {
             'test': entity_mock
         }
-        ds = Dataset(resource)
+        ds = MutableDataset(resource)
         with pytest.raises(ValueError) as err:
             ds.combine_categorical('unknown', CATEGORY_MAP, CATEGORY_NAMES, name='name', alias='alias')
 
@@ -153,7 +154,7 @@ class TestCombine(TestCase):
         resource.variables.by.return_value = {
             'test': entity_mock,
         }
-        ds = Dataset(resource)
+        ds = MutableDataset(resource)
         with pytest.raises(ValueError) as err:
             ds.combine_categorical('test', CATEGORY_MAP, CATEGORY_NAMES, name='name', alias='alias')
         ds.resource.variables.create.assert_called_with(RECODES_PAYLOAD)
@@ -173,7 +174,7 @@ class TestCombine(TestCase):
         tuple_mock.entity.self = var_url
 
         entity = Variable(tuple_mock, resource)
-        ds = Dataset(resource)
+        ds = MutableDataset(resource)
         with pytest.raises(ValueError) as err:
             ds.combine_categorical(entity, CATEGORY_MAP, CATEGORY_NAMES, name='name', alias='alias')
         ds.resource.variables.create.assert_called_with(RECODES_PAYLOAD)
@@ -203,7 +204,7 @@ class TestCombine(TestCase):
             'test': entity_mock
         }
 
-        ds = Dataset(resource)
+        ds = MutableDataset(resource)
         with pytest.raises(ValueError) as err:
             ds.combine_multiple_response('test', RESPONSE_MAP, RESPONSE_NAMES, name='name', alias='alias')
 
@@ -233,7 +234,7 @@ class TestCombine(TestCase):
         }
 
         # make the actual response call
-        ds = Dataset(resource)
+        ds = MutableDataset(resource)
         with pytest.raises(ValueError) as err:
             ds.combine_multiple_response('test', RESPONSE_MAP, RESPONSE_NAMES, name='name', alias='alias')
         resource.variables.create.assert_called_with(COMBINE_RESPONSES_PAYLOAD)
@@ -265,7 +266,7 @@ class TestCombine(TestCase):
             'test': entity_mock
         }
 
-        ds = Dataset(resource)
+        ds = MutableDataset(resource)
 
         with pytest.raises(ValueError) as err:
             ds.combine_multiple_response(entity_mock, RESPONSE_MAP, RESPONSE_NAMES, name='name', alias='alias')
@@ -275,7 +276,7 @@ class TestCombine(TestCase):
 
 class TestRecode(TestCase):
 
-    @mock.patch('scrunch.datasets.get_dataset')
+    @mock.patch('scrunch.mutable_dataset.get_dataset')
     def test_recode_categoricals(self, get_dataset_mock):
         categories = [
             {
@@ -333,7 +334,7 @@ class TestRecode(TestCase):
         ds_res = mock.MagicMock()
         ds_res.self = dataset_url
         ds_res.follow.return_value = table_mock
-        dataset = Dataset(ds_res)
+        dataset = MutableDataset(ds_res)
         dataset.create_categorical([
             {'id': 1, 'name': 'Straight', 'case': 'sexuality.any([1])'},
             {'id': 2, 'name': 'LGBTQ+', 'case': 'sexuality.any([2, 3, 4, 5])'}
@@ -377,7 +378,7 @@ class TestRecode(TestCase):
             }
         })
 
-    @mock.patch('scrunch.datasets.get_dataset')
+    @mock.patch('scrunch.mutable_dataset.get_dataset')
     def test_recode_multiple_responses(self, get_dataset_mock):
         dataset_id = '123'
         categories = [
@@ -464,7 +465,7 @@ class TestRecode(TestCase):
         Variable(var_res, ds_res)
         ds_res.self = dataset_url
         ds_res.follow.return_value = table_mock
-        dataset = Dataset(ds_res)
+        dataset = MutableDataset(ds_res)
         subvar_mock = mock.MagicMock()
         subvar_mock.self = var_url
         subvar_mock.id = 'subvar'
